Add new file for maintenance logic
